+++ src/umpire/strategy/DynamicPool.cpp
@@ -33,6 +33,7 @@ DynamicPool::DynamicPool(const std::string& name,
                          const int align_bytes) noexcept :
   AllocationStrategy(name, id),
   m_allocator{allocator.getAllocationStrategy()},
+  m_initial_alloc_bytes{round_up(initial_alloc_bytes, align_bytes)},
   m_min_alloc_bytes{round_up(min_alloc_bytes, align_bytes)},
   m_align_bytes{align_bytes},
   m_coalesce_heuristic{coalesce_heuristic},
@@ -78,7 +79,12 @@ DynamicPool::~DynamicPool()
     std::size_t whole_bytes;
     std::tie(addr, is_head, whole_bytes) = rec.second;
     // Deallocate if this is a whole block
-    if (is_head && bytes == whole_bytes) m_allocator->deallocate(addr);
+    if (is_head && bytes == whole_bytes) deallocateBlock(addr, bytes);
+  }
+
+  if (m_used_map.size() == 0) {
+    UMPIRE_ASSERT(m_curr_bytes == 0);
+    UMPIRE_ASSERT(m_actual_bytes == 0);
   }
 }
 
@@ -114,7 +120,7 @@ DynamicPool::findFreeBlock(std::size_t bytes) const
   return iter;
 }
 
-void* DynamicPool::allocateFromResource(std::size_t bytes)
+void* DynamicPool::allocateBlock(std::size_t bytes)
 {
   void* ptr{nullptr};
   try {
@@ -127,6 +133,7 @@ void* DynamicPool::allocateFromResource(std::size_t bytes)
                << getFreeBlocks() << " Free Blocks, "
                << getInUseBlocks() << " Used Blocks\n"
       );
+    mergeFreeBlocks();
     releaseFreeBlocks();
     UMPIRE_LOG(Error,
                "\n\tMemory exhausted at allocation resource.  "
@@ -152,7 +159,17 @@ void* DynamicPool::allocateFromResource(std::size_t bytes)
       throw;
     }
   }
+
+  // Add to count
+  m_actual_bytes += bytes;
+
   return ptr;
+}
+
+void DynamicPool::deallocateBlock(void* ptr, std::size_t bytes)
+{
+  m_actual_bytes -= bytes;
+  m_allocator->deallocate(ptr);
 }
 
 void* DynamicPool::allocate(std::size_t bytes)
@@ -175,7 +192,7 @@ void* DynamicPool::allocate(std::size_t bytes)
     insertUsed(ptr, rounded_bytes, is_head, whole_bytes);
 
     // Remove the entry from the free map
-    const std::size_t free_size = iter->first;
+    const std::size_t free_size{iter->first};
     m_free_map.erase(iter);
 
     m_curr_bytes += rounded_bytes;
@@ -188,12 +205,14 @@ void* DynamicPool::allocate(std::size_t bytes)
     }
   } else {
     // Allocate new block -- note this does not check whether alignment is met
-    const std::size_t alloc_bytes{std::max(rounded_bytes, m_min_alloc_bytes)};
-    ptr = allocateFromResource(alloc_bytes);
+    const std::size_t min_block_size =
+      ( m_actual_bytes == 0 ) ? m_initial_alloc_bytes : m_min_alloc_bytes;
+
+    const std::size_t alloc_bytes{std::max(rounded_bytes, min_block_size)};
+    ptr = allocateBlock(alloc_bytes);
 
     // Add used
     insertUsed(ptr, rounded_bytes, true, alloc_bytes);
-    m_actual_bytes += alloc_bytes;
     m_curr_bytes += rounded_bytes;
 
     const std::size_t left_bytes{alloc_bytes - rounded_bytes};
@@ -227,13 +246,16 @@ void DynamicPool::deallocate(void* ptr)
     // Insert in free map
     insertFree(ptr, bytes, is_head, whole_bytes);
 
-    // remove from used map
+    // Remove from used map
     m_used_map.erase(iter);
 
+    // Update currentSize
     m_curr_bytes -= bytes;
   } else {
     UMPIRE_ERROR("Cound not found ptr = " << ptr);
   }
+
+  mergeFreeBlocks();
 
   if (m_coalesce_heuristic(*this)) {
     UMPIRE_LOG(Debug, this
@@ -303,6 +325,8 @@ void DynamicPool::mergeFreeBlocks()
 
   using PointerMap = std::map<Pointer, SizeTuple>;
 
+  UMPIRE_LOG(Debug, "() Free blocks before: " << getFreeBlocks());
+
   // Make a free block map from pointers -> size pairs
   PointerMap free_pointer_map;
 
@@ -337,6 +361,7 @@ void DynamicPool::mergeFreeBlocks()
     // Check if we can merge *it and *next_it
     const bool contiguous{this_addr + this_bytes == next_addr};
     if (contiguous && !next_is_head) {
+      UMPIRE_ASSERT(this_whole_bytes == next_whole_bytes);
       std::get<0>(it->second) += next_bytes;
       next_it = free_pointer_map.erase(next_it);
     } else {
@@ -354,14 +379,13 @@ void DynamicPool::mergeFreeBlocks()
     std::tie(bytes, is_head, whole_bytes) = rec.second;
     insertFree(ptr, bytes, is_head, whole_bytes);
   }
+
+  UMPIRE_LOG(Debug, "() Free blocks after: " << getFreeBlocks());
 }
 
 std::size_t DynamicPool::releaseFreeBlocks()
 {
   UMPIRE_LOG(Debug, "()");
-
-  // Coalesce first so that we are able to release the most memory possible
-  mergeFreeBlocks();
 
   std::size_t released_bytes{0};
 
@@ -376,8 +400,7 @@ std::size_t DynamicPool::releaseFreeBlocks()
     std::tie(ptr, is_head, whole_bytes) = it->second;
     if (is_head && bytes == whole_bytes) {
       released_bytes += bytes;
-      m_actual_bytes -= bytes;
-      m_allocator->deallocate(ptr);
+      deallocateBlock(ptr, bytes);
       it = m_free_map.erase(it);
     } else {
       ++it;
@@ -400,15 +423,10 @@ void DynamicPool::coalesce()
     const std::size_t released_bytes{releaseFreeBlocks()};
     // Deallocated and removed released_bytes from m_free_map
 
-    // If this removed anything from the map, re-allocate a single large chunk
+    // If this removed anything from the map, re-allocate a single large chunk and insert to free map
     if (released_bytes > 0) {
-      const std::size_t actual_bytes{round_up(released_bytes, m_align_bytes)};
-
-      const Pointer ptr{m_allocator->allocate(actual_bytes)};
-      m_actual_bytes += actual_bytes;
-
-      // Add used
-      insertFree(ptr, actual_bytes, true, actual_bytes);
+      const Pointer ptr{allocateBlock(released_bytes)};
+      insertFree(ptr, released_bytes, true, released_bytes);
     }
   }
 }